chore: move mojoTimeDeltaToJSDate into mojomUtils
