+++ components/brave_wallet_ui/utils/datetime-utils.test.ts
@@ -1,18 +1,4 @@
-import { mojoTimeDeltaToJSDate, formatDateAsRelative } from './datetime-utils'
-
-describe('Convertting from MojoTime works', () => {
-  test('0 epoch time value', () => {
-    // Epoch time
-    expect(mojoTimeDeltaToJSDate({ microseconds: 0 })).toEqual(new Date(null))
-  })
-  test('converts normal values ok', () => {
-    const milliseconds = 1635363887716
-    const microseconds = milliseconds * 1000
-    const expectedDate = new Date()
-    expectedDate.setTime(milliseconds)
-    expect(mojoTimeDeltaToJSDate({ microseconds })).toEqual(expectedDate)
-  })
-})
+import { formatDateAsRelative } from './datetime-utils'
 
 describe('Relative dates format correctly', () => {
   const now = new Date(2021, 2, 1)
